Refactor guildclub into multiple cogs
doing the same thing as guildsignal
